requirements.txt
discord.py
python-dotenv
youtube_dl